refactor/split tests
